Fix migrate_notifications.py script
