+++ migrate_notifications.py
@@ -1,7 +1,9 @@
 import os
 import argparse
 import fetchnotifications as fetchnotifications
+import fetchworkflows as fetchworkflows
 import library.clients.notificationsclient as notificationsclient
+import library.clients.workflowsclient as workflowsclient
 import library.localstore as store
 import library.migrationlogger as m_logger
 import library.utils as utils
@@ -9,6 +11,7 @@ import library.utils as utils
 
 log = m_logger.get_logger(os.path.basename(__file__))
 nc = notificationsclient.NotificationsClient()
+wc = workflowsclient.WorkflowsClient()
 
 
 def print_args(args, src_api_key, src_region, tgt_api_key, tgt_region):
@@ -30,8 +33,8 @@ def configure_parser():
     parser.add_argument('--targetApiKey', nargs=1, type=str, required=True, help='Target API Key, \
                                                                     or set environment variable ENV_TARGET_API_KEY')
     parser.add_argument('--targetRegion', type=str, nargs=1, required=False, help='targetRegion us(default) or eu')
-    parser.add_argument('--destinations', dest='destinations', required=False, action='store_true', help='Migrate destinations')
-    parser.add_argument('--channels', dest='channels', required=False, action='store_true', help='Migrate channels')
+    # parser.add_argument('--destinations', dest='destinations', required=False, action='store_true', help='Migrate destinations')
+    # parser.add_argument('--channels', dest='channels', required=False, action='store_true', help='Migrate channels')
     return parser
 
 
@@ -51,6 +54,12 @@ def create_channel(channel, tgt_acct, tgt_api_key, tgt_region):
         log.info(f"Created channel: {channel['name']} of type {channel['type']}")
     else:
         log.warn(f"Unsupported channel type: {channel['type']}, for channel: {channel['name']}")
+
+
+def create_workflow(workflow, tgt_acct, tgt_api_key, tgt_region):
+    log.info(f"Creating workflow: {workflow['name']}")
+    wc.create_workflow(workflow, tgt_api_key, tgt_acct, tgt_region)
+    log.info(f"Created workflow: {workflow['name']}")
 
 
 def migrate_destinations(src_acct, src_api_key, src_region, tgt_acct, tgt_api_key, tgt_region):
@@ -82,6 +91,82 @@ def migrate_channels(src_acct, src_api_key, src_region, tgt_acct, tgt_api_key, t
     return channels_by_source_id
 
 
+def migrate_workflows(src_acct, src_api_key, src_region, tgt_acct, tgt_api_key, tgt_region, channels_by_source_id, policies_by_source_id):
+    log.info('Workflows migration started.')
+    workflows_by_source_id = fetchworkflows.fetch_workflows(src_api_key, src_acct, src_region)
+    for workflow in workflows_by_source_id.values():
+        hasError = False
+        log.info(f"Workflow name: {workflow['name']}")
+        # Enrich destinationConfigurations with target channel ids
+        log.info(f"Enriching destination configurations for target account: {tgt_acct}")
+        if 'destinationConfigurations' in workflow:
+            # Splice workflow['destinationConfigurations'] to contain only supported destinations
+            workflow['destinationConfigurations'][:] = [destination_configuration for destination_configuration in workflow['destinationConfigurations'] if destination_configuration['type'] in notificationsclient.SUPPORTED_DESTINATIONS]
+            if len(workflow['destinationConfigurations']) < 1:
+                log.warning(f"Workflow name: {workflow['name']} does not contain a supported destination")
+                continue
+            for destination_configuration in workflow['destinationConfigurations']:
+                if 'channelId' in destination_configuration:
+                    source_channel_id = destination_configuration['channelId']
+                    if source_channel_id in channels_by_source_id:
+                        channel = channels_by_source_id.get(source_channel_id)
+                        if 'targetChannelId' in channel:
+                            destination_configuration['targetChannelId'] = channel['targetChannelId']
+                            log.info(f"Target channel id: {destination_configuration['targetChannelId']} found for source channel id: {source_channel_id}")
+                        else:
+                            hasError = True
+                            log.error(f"Unable to create workflow name: {workflow['name']}. Target channel id unavailable for source channel id: {source_channel_id} with type: {channel['type']}")
+                    else:
+                        hasError = True
+                        log.error(f"Unable to create workflow name: {workflow['name']}. Source channel id: {source_channel_id} unavailable")
+        else:
+            hasError = True
+            log.info(f"Workflow name: {workflow['name']} with id: {workflow['id']} has no destinationConfigurations: {workflow}")
+        # Enrich issuesFilter with target account id and source policy ids
+        log.info(f"Enriching issues filter for target account: {tgt_acct}")
+        if "issuesFilter" in workflow:
+            workflow['issuesFilter']['targetAccountId'] = int(tgt_acct)
+            for predicate in workflow['issuesFilter']['predicates']:
+                if predicate['attribute'] == 'labels.policyIds':
+                    targetValues = []
+                    for source_policy_id in predicate['values']:
+                        if int(source_policy_id) in policies_by_source_id:
+                            policy = policies_by_source_id.get(int(source_policy_id))
+                            if 'targetPolicyId' in policy:
+                                targetValues.append(str(policy['targetPolicyId']))
+                                log.info(f"Target policy id: {str(policy['targetPolicyId'])} found for source policy id: {source_policy_id} ")
+                            else:
+                                hasError = True
+                                log.error(f"Unable to create workflow name: {workflow['name']}. Target policy id unavailable for source policy id: {source_policy_id}")
+                        else:
+                            hasError = True
+                            log.error(f"Unable to create workflow name: {workflow['name']}. Target policy id unavailable for source policy id: {source_policy_id}")
+                    if len(targetValues) > 0:
+                        predicate['targetValues'] = targetValues
+                else:
+                    log.debug(f"Ignoring predicate {predicate}")
+        else:
+            hasError = True
+            log.info(f"Workflow name: {workflow['name']} with id: {workflow['id']} has no issuesFilter: {workflow}")
+        # Create the workflow
+        if not hasError:
+            create_workflow(workflow, tgt_acct, tgt_api_key, tgt_region)
+        else:
+            log.error(f"Unable to create workflow name: {workflow['name']}, {workflow}")
+    log.info('Workflows migration complete.')
+    return channels_by_source_id
+
+
+def get_policies_by_source_id(source_account):
+    # Get policies by source id
+    data = store.load_alert_policies(source_account)
+    if (data is None) or ('policies' not in data):
+        log.error('No policies found in local store')
+        utils.error_message_and_exit(f'No policies found for sourceAccount {source_account}. Please ensure you have run store_policies.py, migrateconditions.py, and migratepolicies.py for the source account.')
+    # Create the policies_by_source_id dictionary
+    return {policy['id']: policy for policy in data['policies']}
+
+
 def main():
     parser = configure_parser()
     args = parser.parse_args()
@@ -93,14 +178,16 @@ def main():
         utils.error_and_exit('target_api_key', 'ENV_TARGET_API_KEY')
     src_region = utils.ensure_source_region(args)
     tgt_region = utils.ensure_target_region(args)
+    source_account = str(args.sourceAccount[0])
+    target_account = str(args.targetAccount[0])
+    policies_by_source_id = get_policies_by_source_id(source_account)
     print_args(args, src_api_key, src_region, tgt_api_key, tgt_region)
-    if args.destinations:
-        migrate_destinations(args.sourceAccount[0], src_api_key, src_region, args.targetAccount[0], tgt_api_key, tgt_region)
-    elif args.channels:
-        # TODO missing destinations_by_source_id argument!
-        migrate_channels(args.sourceAccount[0], src_api_key, src_region, args.targetAccount[0], tgt_api_key, tgt_region)
-    else:
-        log.info("pass [--destinations | --channels] to fetch configuration")
+    # Migrate notification destinations
+    destinations_by_source_id = migrate_destinations(source_account, src_api_key, src_region, target_account, tgt_api_key, tgt_region)
+    # Migrate notification channels
+    channels_by_source_id = migrate_channels(source_account, src_api_key, src_region, target_account, tgt_api_key, tgt_region, destinations_by_source_id)
+    # Migrate workflows 
+    workflows_by_source_id = migrate_workflows(source_account, src_api_key, src_region, target_account, tgt_api_key, tgt_region, channels_by_source_id, policies_by_source_id)
 
 
 if __name__ == '__main__':
